Add parameters to Renderer::Draw() OpenGL wrapper functions
Also:
- Simplified calls to OpenGL wrapper functions by adding default parameters
- Fixed C++ vs GL type casting
- Clarified comments

+++ OpenGLProject/OpenGLProject/Rendering/ShaderLoader.h
@@ -12,7 +12,7 @@ class Shader;
 // To be used to refer to any Shader instead of relying on raw strings (layer of security over the existence of LookUpIDs when instantiating or look-up functions)
 namespace ShaderLookUpID
 {
-	static const int Num = 5;
+	static const size_t Num = 5;
 
 	// Enum elements do not correspond to GLSL Shader names but on which Scene Entity/Object Mesh they are applied to
 	enum Enum
